requirements.txt
flask
flask-cors
openpyxl
qrcode
Pillow
gunicorn
gspread
google-auth
